services/storage_account/func_send_excel_mark_delete/requirements.txt
pytest
black
python-dotenv
azure.data.tables
azure.core
azure.identity
azure.mgmt.storage
azure.mgmt.monitor
azure-keyvault-secrets
pandas
pytest-cov
openpyxl
azure.storage.blob